gasneti_sharedLimit: fix overflow on ILP32 targets
This commit adds casts to correct overflow seen when multiplying two
32-bit `unsigned long` values.  The overflow was leading to bogus
  "per-host segment limit [..] is too small..."
errors on ILP32 systems with more than 4GiB of `/dev/shm` available
for POSIX shared memory.

This resolves Bug 4751 - incorrect segment limit computation on ILP32

+++ gasnet_mmap.c
@@ -1407,7 +1407,7 @@ uint64_t gasneti_sharedLimit(void) {
     if (fd >= 0) {
       // TODO: for now we ignore any errors here
       if (0 == fstatvfs(fd, &buf)) {
-        uint64_t free_space = buf.f_bsize * buf.f_bavail;
+        uint64_t free_space = (uint64_t)buf.f_bsize * (uint64_t)buf.f_bavail;
         if (free_space) {
           sharedLimit = MIN(sharedLimit, free_space);
         }
